+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Cinema 4D (C4D)
 - Microsoft Reader (LIT)
+- SolidWorks Assembly (SLDASM)
+- SolidWorks Part (SLDPRT)
 - Standard for the Exchange of Product model data (STEP)
 - Universal Scene Description ASCII (USDA)
 - Universal Scene Description Binary (USDC)